+++ src/FSharp.Formatting.CommandTool/BuildCommand.fs
@@ -13,6 +13,7 @@ open System.Runtime.Serialization
 open System.Runtime.Serialization.Formatters.Binary
 
 open Dotnet.ProjInfo
+open Dotnet.ProjInfo.Workspace
 
 module Utils =
     let saveBinary (object:'T) (fileName:string) =
@@ -44,341 +45,6 @@ module Utils =
             let res = f()
             saveBinary res cacheFile
             res
-
-module Inspect =
-
-    module MSBuild =
-        type MSbuildCli =
-             | Property of string * string
-             | Target of string
-             | Switch of string
-             | Project of string
-
-        let sprintfMsbuildArg a =
-            let quote (s: string) =
-                if s.Contains(" ")
-                then sprintf "\"%s\"" s
-                else s
-
-            match a with
-             | Property (k,"") -> sprintf "\"/p:%s=\"" k
-             | Property (k,v) -> sprintf "/p:%s=%s" k v |> quote
-             | Target t -> sprintf "/t:%s" t |> quote
-             | Switch w -> sprintf "/%s" w
-             | Project w -> w |> quote
-
-        let (|ConditionEquals|_|) (str: string) (arg: string) = 
-            if System.String.Compare(str, arg, System.StringComparison.OrdinalIgnoreCase) = 0
-            then Some() else None
-
-        let (|StringList|_|) (str: string)  = 
-            str.Split([| ';' |], System.StringSplitOptions.RemoveEmptyEntries)
-            |> List.ofArray
-            |> Some
-
-    open MSBuild
-
-    type GetProjectInfoErrors<'T> =
-        | UnexpectedMSBuildResult of string
-        | MSBuildFailed of int * 'T
-        | MSBuildSkippedTarget
-
-    [<RequireQualifiedAccess>]
-    type MSBuildExePath =
-        | Path of string
-        | DotnetMsbuild of dotnetExePath: string
-
-    let disableEnvVar envVarName =
-        let oldEnv =
-            match Environment.GetEnvironmentVariable(envVarName) with
-            | null -> None
-            | s ->
-                Environment.SetEnvironmentVariable(envVarName, null)
-                Some s
-        { new IDisposable with 
-            member x.Dispose() =
-                match oldEnv with
-                | None -> ()
-                | Some s -> Environment.SetEnvironmentVariable(envVarName, s) }
-
-    let msbuild msbuildExe run project args =
-        let exe, beforeArgs =
-            match msbuildExe with
-            | MSBuildExePath.Path path -> path, []
-            | MSBuildExePath.DotnetMsbuild path -> path, ["msbuild"]
-        let msbuildArgs =
-            Project(project) :: args @ [ Switch "nologo"; Switch "verbosity:quiet"]
-            |> List.map (MSBuild.sprintfMsbuildArg)
-    
-        //HACK disable FrameworkPathOverride on msbuild, to make installedNETFrameworks work.
-        //     That env var is used only in .net sdk to workaround missing gac assemblies on unix
-        use disableFrameworkOverrideOnMsbuild =
-            match msbuildExe with
-            | MSBuildExePath.Path _ -> disableEnvVar "FrameworkPathOverride"
-            | MSBuildExePath.DotnetMsbuild _ -> { new IDisposable with member x.Dispose() = () }
-
-        match run exe (beforeArgs @ msbuildArgs) with
-        | 0, x -> Ok x
-        | n, x -> Result.Error (MSBuildFailed (n,x))
-
-    let writeTargetFile log templates targetFileDestPath =
-        // https://github.com/dotnet/cli/issues/5650
-
-        let targetFileTemplate = 
-            """
-    <?xml version="1.0" encoding="utf-8" standalone="no"?>
-    <Project ToolsVersion="14.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-      <PropertyGroup>
-        <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-      </PropertyGroup>
-            """
-            + (templates |> String.concat (System.Environment.NewLine))
-            +
-            """
-    </Project>
-            """
-
-        let targetFileOnDisk =
-            if File.Exists(targetFileDestPath) then
-                try
-                    Some (File.ReadAllText targetFileDestPath)
-                with
-                | _ -> None
-            else
-                None
-
-        let newTargetFile = targetFileTemplate.Trim()
-
-        if targetFileOnDisk <> Some newTargetFile then
-            log (sprintf "writing helper target file in '%s'" targetFileDestPath)
-            File.WriteAllText(targetFileDestPath, newTargetFile)
-
-        Ok targetFileDestPath
-
-    type GetResult =
-         | Properties of (string * string) list
-
-    let getNewTempFilePath suffix =
-        let outFile = System.IO.Path.GetTempFileName()
-        //if File.Exists outFile then File.Delete outFile
-        sprintf "%s.%s" outFile suffix
-
-    let bindSkipped f outFile =
-        if not(File.Exists outFile) then
-            Result.Error MSBuildSkippedTarget
-        else
-            f outFile
-
-    let parsePropertiesOut outFile =
-        let firstAndRest (delim: char) (s: string) =
-            match s.IndexOf(delim) with
-            | -1 -> None
-            | index -> Some(s.Substring(0, index), s.Substring(index + 1))
-
-        let lines =
-            File.ReadAllLines(outFile)
-            |> Array.filter (fun s -> s.Length > 0)
-            |> Array.map (fun s -> match s |> firstAndRest '=' with Some x -> Ok x | None -> Result.Error s)
-            |> List.ofArray
-
-        match lines |> List.partition (function Ok _ -> true | Error _ -> false) with
-        | l, [] ->
-            l
-            |> List.choose (function Ok x -> Some x | Error _ -> None)
-            |> (fun x -> Ok x)
-        | _, err ->
-            err
-            |> List.choose (function Ok _ -> None | Error x -> Some x)
-            |> sprintf "invalid temp file content '%A'"
-            |> (fun x -> Result.Error (UnexpectedMSBuildResult x))
-
-    let getProperties props =
-        let templateF isCrossgen =
-            """
-      <Target Name="_Inspect_GetProperties_""" + (if isCrossgen then "CrossGen" else "NotCrossGen") + """"
-              Condition=" '$(IsCrossTargetingBuild)' """ + (if isCrossgen then "==" else "!=") + """ 'true' "
-              """ + (if isCrossgen then "" else "DependsOnTargets=\"ResolveReferences\"" ) + """ >
-        <ItemGroup>
-            """
-            + (
-                props
-                |> List.mapi (fun i p -> sprintf """
-            <_Inspect_GetProperties_OutLines Include="P%i">
-                <PropertyName>%s</PropertyName>
-                <PropertyValue>$(%s)</PropertyValue>
-            </_Inspect_GetProperties_OutLines>
-                                                 """ i p p)
-                |> List.map (fun s -> s.TrimEnd())
-                |> String.concat (System.Environment.NewLine) )
-            +
-            """
-        </ItemGroup>
-        <Message Text="%(_Inspect_GetProperties_OutLines.PropertyName)=%(_Inspect_GetProperties_OutLines.PropertyValue)" Importance="High" />
-        <WriteLinesToFile
-                Condition=" '$(_Inspect_GetProperties_OutFile)' != '' "
-                File="$(_Inspect_GetProperties_OutFile)"
-                Lines="@(_Inspect_GetProperties_OutLines -> '%(PropertyName)=%(PropertyValue)')"
-                Overwrite="true" 
-                Encoding="UTF-8"/>
-      </Target>
-            """.Trim()
-
-        //doing like that (crossgen/notcrossgen) because ResolveReferences doesnt exists
-        //if is crossgen
-
-        let templateAll =
-            """
-      <Target Name="_Inspect_GetProperties"
-              DependsOnTargets="_Inspect_GetProperties_CrossGen;_Inspect_GetProperties_NotCrossGen" />
-            """
-
-        let template =
-            [ templateF true
-              templateF false
-              templateAll ]
-            |> String.concat (System.Environment.NewLine)
-    
-        let outFile = getNewTempFilePath "GetProperties.txt"
-        let args =
-            [ Target "_Inspect_GetProperties"
-              Property ("_Inspect_GetProperties_OutFile", outFile) ]
-        template, args, (fun () -> outFile
-                                   |> bindSkipped parsePropertiesOut
-                                   |> Result.map Properties)
-
-    let uninstall_old_target_file log (projPath: string) =
-        let projDir, projName = Path.GetDirectoryName(projPath), Path.GetFileName(projPath)
-        let objDir = Path.Combine(projDir, "obj")
-        let targetFileDestPath = Path.Combine(objDir, (sprintf "%s.proj-info.targets" projName))
-
-        log (sprintf "searching deprecated target file in '%s'." targetFileDestPath)
-        if File.Exists targetFileDestPath then
-            log (sprintf "found deprecated target file in '%s', deleting." targetFileDestPath)
-            File.Delete targetFileDestPath
-
-    let getProjectInfos log msbuildExec getters additionalArgs projPath =
-
-        let templates, argsList, parsers = 
-            getters
-            |> List.map (fun getArgs -> getArgs ())
-            |> List.unzip3
-
-        let args = argsList |> List.concat
-
-        // remove deprecated target file, if exists
-        projPath
-        |> uninstall_old_target_file log
-
-        getNewTempFilePath "proj-info.hook.targets"
-        |> writeTargetFile log templates
-        |> Result.bind (fun targetPath -> msbuildExec projPath (args @ additionalArgs @ [ Property("CustomAfterMicrosoftCommonTargets", targetPath); Property("CustomAfterMicrosoftCommonCrossTargetingTargets", targetPath) ]))
-        |> Result.map (fun _ -> parsers |> List.map (fun parse -> parse ()))
-
-module internal InspectSln =
-
-      let private normalizeDirSeparators (path: string) =
-          match System.IO.Path.DirectorySeparatorChar with
-          | '\\' -> path.Replace('/', '\\')
-          | '/' -> path.Replace('\\', '/')
-          | _ -> path
-
-      type SolutionData = {
-            Items: SolutionItem list
-            Configurations: SolutionConfiguration list
-            }
-      and SolutionConfiguration = {
-            Id: string
-            ConfigurationName: string
-            PlatformName: string
-            IncludeInBuild: bool
-            }
-      and SolutionItem = {
-            Guid: Guid
-            Name: string
-            Kind: SolutionItemKind
-            }
-      and SolutionItemKind =
-            | MsbuildFormat of SolutionItemMsbuildConfiguration list
-            | Folder of (SolutionItem list) * (string list)
-            | Unsupported
-            | Unknown
-      and SolutionItemMsbuildConfiguration = {
-            Id: string
-            ConfigurationName: string
-            PlatformName: string
-            }
-
-      let tryParseSln (slnFilePath: string) = 
-        let parseSln (sln: Microsoft.Build.Construction.SolutionFile) =
-            let slnDir = Path.GetDirectoryName slnFilePath
-            let makeAbsoluteFromSlnDir =
-                let makeAbs (path: string) =
-                    if Path.IsPathRooted path then
-                        path
-                    else
-                        Path.Combine(slnDir, path)
-                        |> Path.GetFullPath
-                normalizeDirSeparators >> makeAbs
-            let rec parseItem (item: Microsoft.Build.Construction.ProjectInSolution) =
-                let parseKind (item: Microsoft.Build.Construction.ProjectInSolution) =
-                    match item.ProjectType with
-                    | Microsoft.Build.Construction.SolutionProjectType.KnownToBeMSBuildFormat ->
-                        (item.RelativePath |> makeAbsoluteFromSlnDir), SolutionItemKind.MsbuildFormat []
-                    | Microsoft.Build.Construction.SolutionProjectType.SolutionFolder ->
-                        let children =
-                            sln.ProjectsInOrder
-                            |> Seq.filter (fun x -> x.ParentProjectGuid = item.ProjectGuid)
-                            |> Seq.map parseItem
-                            |> List.ofSeq
-                        let files =
-                            item.FolderFiles
-                            |> Seq.map makeAbsoluteFromSlnDir
-                            |> List.ofSeq
-                        item.ProjectName, SolutionItemKind.Folder (children, files)
-                    | Microsoft.Build.Construction.SolutionProjectType.EtpSubProject
-                    | Microsoft.Build.Construction.SolutionProjectType.WebDeploymentProject
-                    | Microsoft.Build.Construction.SolutionProjectType.WebProject ->
-                        (item.ProjectName |> makeAbsoluteFromSlnDir), SolutionItemKind.Unsupported
-                    | Microsoft.Build.Construction.SolutionProjectType.Unknown
-                    | _ ->
-                        (item.ProjectName |> makeAbsoluteFromSlnDir), SolutionItemKind.Unknown
-
-                let name, itemKind = parseKind item 
-                { Guid = item.ProjectGuid |> Guid.Parse
-                  Name = name
-                  Kind = itemKind }
-
-            let items =
-                sln.ProjectsInOrder
-                |> Seq.filter (fun x -> isNull x.ParentProjectGuid)
-                |> Seq.map parseItem
-            let data = {
-                Items = items |> List.ofSeq
-                Configurations = []
-            }
-            (slnFilePath, data)
-
-        try
-            slnFilePath
-            |> Microsoft.Build.Construction.SolutionFile.Parse
-            |> parseSln
-            |> Choice1Of2
-        with ex ->
-            Choice2Of2 ex
-
-      let loadingBuildOrder (data: SolutionData) =
-
-        let rec projs (item: SolutionItem) =
-            match item.Kind with
-            | MsbuildFormat items ->
-                [ item.Name ]
-            | Folder (items, _) ->
-                items |> List.collect projs
-            | Unsupported
-            | Unknown ->
-                []
-
-        data.Items |> List.collect projs
 
 module Crack =
 
@@ -482,17 +148,18 @@ module Crack =
                    PackageReleaseNotes = props |> Map.tryFind "PackageReleaseNotes"
                    RepositoryCommit = props |> Map.tryFind "RepositoryCommit" |}
                 
+            | Ok ok -> failwithf "huh? ok = %A" ok
             | Error err -> failwithf "error - %s\nlog - %s" (err.ToString()) (String.concat "\n" msgs)
         | Ok ok -> failwithf "huh? ok = %A" ok
         | Error err -> failwithf "error - %s\nlog - %s" (err.ToString()) (String.concat "\n" msgs)
                 
     let getProjectsFromSlnFile (slnPath : string) =
         match InspectSln.tryParseSln slnPath with
-        | Choice1Of2 (_, slnData) ->
+        | Ok (_, slnData) ->
             InspectSln.loadingBuildOrder slnData
 
             //this.LoadProjects(projs, crosstargetingStrategy, useBinaryLogger, numberOfThreads)
-        | Choice2Of2 d ->
+        | Error d ->
             failwithf "cannot load the sln: %A" d
 
 type CoreBuildOptions(watch) =
@@ -525,6 +192,9 @@ type CoreBuildOptions(watch) =
 
     [<Option("parameters", Required = false, HelpText = "Additional substitution parameters for templates.")>]
     member val parameters = Seq.empty<string> with get, set
+
+    [<Option("clean", Required = false, Default=false, HelpText = "Clean the output directory.")>]
+    member val clean = false with get, set
 
     member x.Execute() =
         let mutable res = 0
@@ -710,6 +380,19 @@ type CoreBuildOptions(watch) =
                         queued <- false; run()) ) 
             watcher.EnableRaisingEvents <- true
             printfn "Building docs first time..." 
+
+        if x.clean then
+            let rec clean dir =
+                for file in Directory.EnumerateFiles dir do
+                    File.Delete file |> ignore
+                for subdir in Directory.EnumerateDirectories dir do
+                   if not (subdir.StartsWith ".") then
+                       clean subdir
+            if x.output <> "/" && x.output <> "." && Path.GetFullPath x.output <> Path.GetFullPath x.input && not (String.IsNullOrEmpty x.output) then
+                try clean (Path.GetFullPath x.output)
+                with e -> printfn "warning: error during cleaning, continuing: %s" e.Message
+            else
+                printfn "warning: skipping cleaning due to strange output path: \"%s\"" x.output
 
         lock monitor run
         queued <- false